Implement common-models crate
And refactor feature-flags to use the common methods.
