+++ rust/Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "common/metrics",
     "common/symbol_data",
     "common/types",
+    "common/models",
     "feature-flags",
     "hook-api",
     "hook-common",
