+++ rust/common/models/src/lib.rs
@@ -1,11 +1,6 @@
-pub mod api;
-pub mod cohorts;
-pub mod config;
-pub mod flags;
-pub mod metrics;
-pub mod properties;
-pub mod router;
-pub mod server;
+pub mod errors;
+pub mod team_models;
+pub mod team_operations;
 
 // Test modules don't need to be compiled with main binary
 // #[cfg(test)]
@@ -13,4 +8,4 @@ pub mod server;
 // or make it a separate feature using cfg(feature = "integration-tests")
 // and then use this feature only in tests.
 // For now, ok to just include in binary
-pub mod utils;
+pub mod test_utils;
