+++ rust/feature-flags/Cargo.toml
@@ -13,6 +13,7 @@ axum-client-ip = { workspace = true }
 chrono = { workspace = true }
 common-cookieless = { path = "../common/cookieless" }
 common-database = { path = "../common/database" }
+common-models = { path = "../common/models" }
 common-redis = { path = "../common/redis" }
 common-types = { path = "../common/types" }
 envconfig = { workspace = true }
@@ -23,11 +24,6 @@ tracing-subscriber = { workspace = true, features = ["env-filter"] }
 bytes = { workspace = true }
 once_cell = "1.18.0"
 rand = { workspace = true }
-redis = { version = "0.23.3", features = [
-    "tokio-comp",
-    "cluster",
-    "cluster-async",
-] }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
@@ -48,7 +44,6 @@ tower-http = { workspace = true }
 petgraph = "0.6.5"
 moka = { workspace = true }
 dateparser = "0.2.1"
-rust_decimal = { workspace = true }
 rayon = "1.10.0"
 percent-encoding = "2.3.1"
