--- rust/feature-flags/src/team/mod.rs
@@ -1,2 +1,0 @@
-pub mod team_models;
-pub mod team_operations;
